start to remove some cruft
